feat: move `area_same_color` to `neighbor` module
